Map moved to homepage(search.js)
Instead of rendering on App.js, map should render on the homepage - which for us is the search.js component.
