+++ client/src/App.js
@@ -1,12 +1,12 @@
 import './App.css';
 import React, { useState } from "react"
-import LoginButton from './components/login';
-import LogoutButton from './components/logout';
-import Profile from './components/profile';
+// import LoginButton from './components/login';
+// import LogoutButton from './components/logout';
+// import Profile from './components/profile';
 import Header from "./components/header";
 import Footer from "./components/footer";
-import Search from "./components/footer";
-import * as parksData from "./data/skateboard-parks.json"
+import Search from "./components/search";
+// import * as parksData from "./data/skateboard-parks.json";
 
 import {
   ApolloProvider,
@@ -15,40 +15,40 @@ import {
   createHttpLink
 } from '@apollo/react-hooks';
 
-// google react api libraries
-import {
-  GoogleMap,
-  useLoadScript,
-  Marker,
-  // InfoWindow,
-} from '@react-google-maps/api';
+// // google react api libraries
+// import {
+//   GoogleMap,
+//   useLoadScript,
+//   Marker,
+//   // InfoWindow,
+// } from '@react-google-maps/api';
 
 // import usePlacesAutocomplete, {
 //   getGeocode,
 //   getLatLng, } 
 //   from 'use-places-autocomplete';
 
-// styling for the map
-import mapStyles from './mapStyles';
+// // styling for the map
+// import mapStyles from './mapStyles';
 
-// set up options to pass to the googlemaps component (to avoid rerendering)
-const libraries = ['places'];
-const mapContainerStyle = {
-  width: "100vw",
-  height: "100vh"
-};
+// // set up options to pass to the googlemaps component (to avoid rerendering)
+// const libraries = ['places'];
+// const mapContainerStyle = {
+//   width: "100vw",
+//   height: "100vh"
+// };
 
-// lat/lng for ottawa, toronto is lat: 43.6532, lng:-79.3831
-const center = {
-  lat: 45.4247,
-  lng: -75.6950
-};
+// // lat/lng for ottawa, toronto is lat: 43.6532, lng:-79.3831
+// const center = {
+//   lat: 45.4247,
+//   lng: -75.6950
+// };
 
-const options = {
-  styles: mapStyles,
-  disableDefaultUI: true,
-  zoomControl: true,
-}
+// const options = {
+//   styles: mapStyles,
+//   disableDefaultUI: true,
+//   zoomControl: true,
+// }
 
 // we need to establish the connection to the back-end server's /graphql endpoint. establish a new link to the GraphQL server at its /graphql endpoint with createHttpLink()
 const httpLink = createHttpLink({
@@ -62,11 +62,11 @@ const client = new ApolloClient({
 });
 
 function App() {
-  // use loadscript hook to setup load script/communicate with google api
-  const { loadError, isLoaded } = useLoadScript({
-    googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAPS_API_KEY,
-    libraries
-  });
+  // // use loadscript hook to setup load script/communicate with google api
+  // const { loadError, isLoaded } = useLoadScript({
+  //   googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAPS_API_KEY,
+  //   libraries
+  // });
   const [currentTab, setCurrentTab] = useState("search");
   const renderTab = () => {
     switch (currentTab) {
@@ -85,8 +85,8 @@ function App() {
     }
   }
 
-  if(loadError) return "Error loading map";
-  if(!isLoaded) return "Loading Maps";
+  // if (loadError) return "Error loading map";
+  // if (!isLoaded) return "Loading Maps";
 
   return (
 
@@ -98,25 +98,7 @@ function App() {
         </Header>
       </div>
       <main>{renderTab()}
-        <div className="map-container">
-          <GoogleMap
-            mapContainerStyle={mapContainerStyle}
-            zoom={10}
-            center={center}
-            options={options}
-          >
-            {/* embbed markers inside maps component */}
-            {parksData.features.map((park) => (
-              <Marker
-                key={park.properties.PARK_ID}
-                position={{
-                  lat: park.geometry.coordinates[1],
-                  lng: park.geometry.coordinates[0]
-                }}
-              />
-            ))}
-          </GoogleMap>
-        </div>
+       
       </main>
       <div>
         <Footer></Footer>
